+++ application/controllers/Info.php
@@ -1,14 +1,20 @@
 <?php
 
-defined('BASEPATH') OR exit('No direct script access allowed');
-
+/*
+	This is a RESTish service controller which is useful for testing.
+	The user receives the model data in JSON format on deman
+	Ex. the URL /info/category will return all data inside the category model.
+*/
 class Info extends Application
 {
+
+	//Echos the name of the scenario in question
 	public function index()
 	{
 		echo("Scenario: Best Warrior");
 	}
 
+	//Return the designated category; however, if none is specified than all  category data will be returned
 	public function category($key = null)
 	{
 		$categories = $this->categories->all();
@@ -29,6 +35,7 @@ class Info extends Application
 		}
 	}
 
+	//Return the designated catalog; however, if none is specified than all catalog data will be returned
 	public function catalog($key = null)
 	{
 		$accessories = $this->accessories->all();
@@ -49,6 +56,7 @@ class Info extends Application
 		}
 	}
 
+	//Return the designated set; however, if none is specified than all set data will be returned
 	public function bundle($key = null)
 	{
 		$set = $this->set->all();